Support generating a calendar with a 3×4 month grid
